#29: Added `.matches(_:ignoring:)`
This allows two semantic versions to be compared without regarding certain identifiers
- Also moved `Comparable` conformance to a new file; placed new code there
- Also moved `Comparable` tests to a new file; placed new tests there
